Added stone texture, updated dirt texture. Updated interface. Probably some other things too.

+++ Source/Source/Palette.h
@@ -28,7 +28,7 @@ private:
 
 	BitArray data_;
 	std::vector<PaletteEntry> palette_;
-	size_t paletteEntryLength_ = 1;
+	unsigned paletteEntryLength_ = 1;
 };
